requirements.txt
torch
torchvision
flask
numpy
notebook
python-box
pyYaml
joblib
types-pyYaml
Flask
Flask-Cors
tqdm
ensure
torch
Pillow
scikit-learn
vaderSentiment
nltk

-e